+++ test-crates/pyo3-mixed-py-subdir/python/pyo3_mixed_py_subdir/__init__.py
@@ -1,0 +1,6 @@
+from .python_module.double import double
+from .pyo3_mixed_py_subdir import get_21
+
+
+def get_42() -> int:
+    return double(get_21)